requirements.txt
transformers 
torch 
python-dotenv 
streamlit 
huggingface-hub 
